Remove `PatientResource` and its associated `Create`, `Edit`, and `List` pages, refactoring and consolidating `Patient` management logic within Filament structure.
