--- app/Filament/Clusters/Managment/Resources/PatientResource/Pages/EditPatient.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Filament\Clusters\Managment\Resources\PatientResource\Pages;
-
-use App\Filament\Clusters\Managment\Resources\PatientResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditPatient extends EditRecord
-{
-    protected static string $resource = PatientResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\DeleteAction::make(),
-        ];
-    }
-}
-
