--- app/Filament/Clusters/Managment/Resources/PatientResource/Pages/ListPatients.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Filament\Clusters\Managment\Resources\PatientResource\Pages;
-
-use App\Filament\Clusters\Managment\Resources\PatientResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-class ListPatients extends ListRecords
-{
-    protected static string $resource = PatientResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
-
