+++ app/Filament/Resources/PatientResource/Pages/CreatePatient.php
@@ -3,10 +3,10 @@
 namespace App\Filament\Resources\PatientResource\Pages;
 
 use App\Filament\Resources\PatientResource;
-use Filament\Actions;
 use Filament\Resources\Pages\CreateRecord;
 
 class CreatePatient extends CreateRecord
 {
     protected static string $resource = PatientResource::class;
 }
+
