+++ app/Filament/Resources/PatientResource/Pages/EditPatient.php
@@ -17,3 +17,4 @@ class EditPatient extends EditRecord
         ];
     }
 }
+
