+++ app/Filament/Resources/PatientResource/Pages/ListPatients.php
@@ -17,3 +17,4 @@ class ListPatients extends ListRecords
         ];
     }
 }
+
